+++ visualization/app/codeCharta/ui/toolBar/toolBar.module.ts
@@ -10,6 +10,7 @@ import { GlobalConfigurationButtonComponent } from "./globalConfigurationButton/
 import { HoveredNodePathPanelComponent } from "./hoveredNodePathPanel/hoveredNodePathPanel.component"
 import { FilePanelComponent } from "../filePanel/filePanel.component"
 import { ScreenshotButtonComponent } from "../screenshotButton/screenshotButton.component"
+import { DownloadButtonComponent } from "./downloadButton/downloadButton.component"
 
 angular
 	.module("app.codeCharta.ui.toolBar", ["app.codeCharta.state", "app.codeCharta.ui.dialog"])
@@ -20,3 +21,4 @@ angular
 	.directive("ccHoveredNodePathPanel", downgradeComponent({ component: HoveredNodePathPanelComponent }))
 	.directive("ccScreenshotButton", downgradeComponent({ component: ScreenshotButtonComponent }))
 	.directive("ccFilePanel", downgradeComponent({ component: FilePanelComponent }))
+	.directive("ccDownloadButton", downgradeComponent({ component: DownloadButtonComponent }))